Tests fixed and moved to a separate project
